[#15] Rename EditVC to ShowDetailVC and Refactor ShowDetailVC
- EditVC는 따로 만들기로 하고, habit 상세 내용만 보여주도록 하고자 이름변경
- 뷰모델 역할 ShowDetailViewModel로 이전
